# api/firewall/openapi.yaml
openapi: 3.0.3

info:
  title: CasaOS Firewall API
  version: v2
  description: |-
    <picture>
        <source media="(prefers-color-scheme: dark)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_dark_night_800px.png">
        <source media="(prefers-color-scheme: light)" srcset="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
        <img alt="CasaOS" src="https://raw.githubusercontent.com/IceWhaleTech/logo/main/casaos/casaos_banner_twilight_blue_800px.png">
    </picture>

servers:
    - url: /v2/firewall

tags:
  - name: Firewall methods
    description: |-
      Low-level API

x-tagGroups:
  - name: Firewall
    tags:
      - Firewall methods

security:
  - access_token: []

paths:
  /firewall:
    get:
      summary: Get State
      description: Get firewall state
      operationId: getState
      tags:
        - Firewall methods
      responses:
        "200":
          $ref: "#/components/responses/GetStateResponseOK"
        "503":
          $ref: "#/components/responses/ResponseServiceUnavailable"

  /port:
    get:
      summary: Get Opened Ports
      description: |-
        (TODO)
      operationId: getOpenedPorts
      tags:
        - Firewall methods
      responses:
        "200":
          $ref: "#/components/responses/GetPortsResponseOK"
        "503":
          $ref: "#/components/responses/ResponseServiceUnavailable"

    post:
      summary: Open  a Port
      description: |-
        (TODO)
      operationId: openPort
      tags:
        - Firewall methods
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Port"
      responses:
        "200":
          $ref: "#/components/responses/ChangePortResponseOK"
        "403":
          $ref: "#/components/responses/ResponseBadRequest"
        "503":
          $ref: "#/components/responses/ResponseServiceUnavailable"
    
    delete:
      summary: Close a Port 
      description: |-
        (TODO)
      operationId: deletePortRule
      tags:
        - Firewall methods
    
    put:
      summary: Update a Port Rule
      description: |-
        (TODO)
      operationId: updatePortRule
      tags:
        - Firewall methods

components:
  securitySchemes:
    access_token:
      type: apiKey
      in: header
      name: Authorization

  responses:
    GetStateResponseOK:
      description: OK
      content:
        application/json:
          schema:
            readOnly: true
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - type: object
                properties:
                  data:
                    type: boolean
                    example: true  
    GetVersionResponseOK:
      description: OK
      content:
        application/json:
          schema:
            readOnly: true
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - type: object
                properties:
                  data:
                    type: string
                    example: "0.4.5" 
    ChangePortResponseOK:
      description: OK
      content:
        application/json:
          schema:
            readOnly: true
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - type: object
                properties:
                  data:
                    type: string
                    example: "success"  
    GetPortsResponseOK:
      description: OK
      content:
        application/json:
          schema:
            readOnly: false
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
              - type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Port"
    ResponseBadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            readOnly: true
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Bad Request"
    ResponseServiceUnavailable:
      description: Service Unavailable
      content:
        application/json:
          schema:
            readOnly: true
            allOf:
              - $ref: "#/components/schemas/BaseResponse"
          example:
            message: "Service Unavailable"
  schemas:
    Rule:
      type: string
   
    Port:
      type: object
      required:
        - port
        - protocol
      properties:
        port:
          type: string
          example: "80"
        protocol:
          type: string
          enum:
            - "tcp"
            - "udp"
        status:
          type: string
          example: "open"

    BaseResponse:
      properties:
        message:
          readOnly: true
          description: message returned by server side if there is any
          type: string
          example: ""